# laravel_project_api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mon API",
        "description": "Description de mon API",
        "contact": {
            "email": "contact@example.com"
        },
        "license": {
            "name": "License MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/products/category/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get products by category",
                "operationId": "078b6dbd2d52eb2efe57ef6bb8c00cb6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of products by category",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Products"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Products": {
                "title": "Products",
                "description": "Products model",
                "properties": {
                    "id": {
                        "description": "The unique identifier for a product",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "The name of the product",
                        "type": "string",
                        "example": "Smartphone"
                    },
                    "description": {
                        "description": "The description of the product",
                        "type": "string",
                        "example": "High-end smartphone with advanced features"
                    },
                    "price": {
                        "description": "The price of the product",
                        "type": "number",
                        "format": "float",
                        "example": 999.99
                    },
                    "stock": {
                        "description": "The stock quantity of the product",
                        "type": "integer",
                        "example": 100
                    },
                    "image": {
                        "description": "The image URL of the product",
                        "type": "string",
                        "example": "https://example.com/image.jpg"
                    }
                },
                "type": "object"
            },
...
        }
    }
}